2dcontext/hit-regions/addHitRegion.nullIDControl-manual.html


<!DOCTYPE html>
<html lang="en">
<head>
<meta charset="UTF-8">
<title>addHitRegion throws NotSupportedError when dictionary object id is NULL and the Control is Null</title>
    <script src="//w3c-test.org/resources/testharness.js"></script>
    <script src="//w3c-test.org/resources/testharnessreport.js"></script>
    <link rel="author" title="Mark Sadecki">
    <link rel="help" href="http://www.w3.org/TR/2dcontext/#dom-context-2d-addhitregion">
</head>
<body>
<div id="log"></div>
<div><canvas id="canvas"><button id="button">button</button></canvas></div>

<script>
test(function () {
    assert_throws({ name: "NotSupportedError" },
        function () {
            var canvas = document.getElementById("canvas");
            // Reset context
            canvas.width = canvas.width;
            var button = document.getElementById("button");
            var context = canvas.getContext("2d");

            // If the arguments object's id member is null and the arguments object's control member is null,
            // throw a NotSupportedError exception and abort these steps.
            // http://www.w3.org/TR/2dcontext/#dom-context-2d-addhitregion
            context.addHitRegion();
        }, "addHitRegion throws NotSupportedError when dictionary object id is NULL and the Control is Null.");
}, "addHitRegion throws NotSupportedError when dictionary object id is NULL and the Control is Null.");
    </script>
    <div id="log"></div>
  </body>
</html>
